+++ benchmark/defect_BENCHMARK.cc
@@ -16,7 +16,7 @@ class DefectBench :
 {
     private:
         unsigned long _size;
-        int _count;
+        unsigned long _count;
     public:
         DefectBench(const std::string & id, unsigned long size, int count) :
             Benchmark(id)
