+++ honei/la/dense_vector_TEST.cc
@@ -181,13 +181,13 @@ class DenseVectorFunctionsTest :
                 }
                 TEST_CHECK_EQUAL(dv.size(), size);
 
-                for (int i=0 ; i<size ; ++i)
+                for (unsigned long i=0 ; i<size ; ++i)
                 {
                     DataType_ s((i+1)/1.23456789);
                     TEST_CHECK_EQUAL_WITHIN_EPS((dv)[i], s,
                         std::numeric_limits<DataType_>::epsilon());
                 }
-                for (int i=0 ; i<size ; ++i)
+                for (unsigned long i=0 ; i<size ; ++i)
                 {
                     DataType_ s((i+5)/1.23456789);
                     (dv)[i] = s;
@@ -215,7 +215,7 @@ class DenseVectorQuickTest :
         virtual void run() const
         {
             DenseVector<DataType_> dv(4711, DataType_(123.987));
-            TEST_CHECK_EQUAL(dv.size(), 4711);
+            TEST_CHECK_EQUAL(dv.size(), 4711ul);
             TEST_CHECK_EQUAL(dv, dv);
             TEST_CHECK_EQUAL_WITHIN_EPS((dv)[4710] , 123.987, sqrt(std::numeric_limits<DataType_>::epsilon()));
             DataType_ s = DataType_(1.2345);
